+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@ CHANGELOG
 ---------
 
 #### Next release
+- New features:
+  - Click anywhere on toolbar to open chapter selection view and move to current book (#206)
 - Update dependencies:
   - Gradle to 7.3
   - Kotlin to 1.6.0